refactor: move LIST_CURRENCIES to constants and update imports in EditUserForm and tests
