+++ app/src/components/forms/EditUserForm.vue
@@ -98,7 +98,8 @@ import { required, minLength } from '@vuelidate/validators'
 import ToolTip from '@/components/ToolTip.vue'
 import { Icon as IconifyIcon } from '@iconify/vue'
 import ButtonUI from '../ButtonUI.vue'
-import { LIST_CURRENCIES, useCurrencyStore } from '@/stores'
+import { useCurrencyStore } from '@/stores'
+import LIST_CURRENCIES from '@/constant'
 import { useClipboard } from '@vueuse/core'
 import { NETWORK } from '@/constant'
 import { ref } from 'vue'
